// src/types/tool.types.ts
export enum EntityType {
    ACCOUNT = 'Account',
    CONTACT = 'Contact',
    DEAL = 'Deal',
    ARTICLE = 'Article',
    CASE = 'Case',
    LEAD = 'Lead'
}